Delete Clerk sign-in and sign-up pages; update link in home page
The commit deletes the sign-in and sign-up pages that were using Clerk's `SignIn` and `SignUp` components. It also updates the link to the sign-in page in the home page to remove the `/auth` prefix. The `publicRoutes` option in the authentication middleware is commented out, meaning that all routes will require authentication.